Add `errored` alert type
Colour was picked by inspecting the build header colour in my browser
which was `background: rgb(245, 166, 35);` and converting that to
hexadecimal.

Note, there is no image for the readme included as part of this commit
as I'm not too sure how to generate it to look consistent with the
others.